Refactor God classes: extract focused responsibilities from 3 oversized classes
DatasetEphemerisSource: extracted binary format parsing, file I/O, and LRU cache
into four dedicated package-private classes (EphemerisV1Parser, BodyFile,
DecodedChunk, LruCache). DatasetEphemerisSource becomes a thin facade.

CMAESTrajectoryOptimizer: extracted CMA-ES execution logic into CMAESRunExecutor
and convergence decisions into AdaptiveConvergenceChecker. Optimizer retains only
the exploration/refinement phase orchestration.

TransfertTwoManeuver: extracted burn-2 resolution and apoapsis detection into
Burn2Resolver. Introduced TransferResult record to eliminate mutable lastXxx
side-effect fields. propagateForOptimization() now returns TransferResult;
TransferTwoManeuverProblem owns lastResult; MissionOptimizer reads from it.
