+++ src/main/java/com/smousseur/orbitlab/simulation/mission/optimizer/CMAESTrajectoryOptimizer.java
@@ -3,14 +3,8 @@ package com.smousseur.orbitlab.simulation.mission.optimizer;
 import com.smousseur.orbitlab.core.OrbitlabException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.hipparchus.analysis.MultivariateFunction;
-import org.hipparchus.optim.*;
-import org.hipparchus.optim.nonlinear.scalar.GoalType;
-import org.hipparchus.optim.nonlinear.scalar.ObjectiveFunction;
-import org.hipparchus.optim.nonlinear.scalar.noderiv.CMAESOptimizer;
 import org.hipparchus.random.MersenneTwister;
 import org.hipparchus.util.FastMath;
-import org.jspecify.annotations.NonNull;
 import org.orekit.propagation.SpacecraftState;
 
 /**
@@ -28,14 +22,9 @@ import org.orekit.propagation.SpacecraftState;
 public class CMAESTrajectoryOptimizer implements TrajectoryOptimizer {
   private static final Logger logger = LogManager.getLogger(CMAESTrajectoryOptimizer.class);
 
-  public static final double EXCEPTION_PENALTY_COST = 1e10;
-
   private final TrajectoryProblem problem;
   private final int maxEvaluations;
   private final int numExplorationRuns;
-  private final double stopFitness;
-  private final double relativeTolerance;
-  private final double absoluteTolerance;
 
   private final MersenneTwister rng = new MersenneTwister();
 
@@ -43,17 +32,13 @@ public class CMAESTrajectoryOptimizer implements TrajectoryOptimizer {
   /** Cost below which we consider we've found a good basin and switch to refinement-only. */
   private static final double GOOD_BASIN_THRESHOLD = 0.01;
 
-  /** Cost above which an exploration run is killed early (bad basin). */
-  private static final double BAD_BASIN_KILL_THRESHOLD = 1.0;
-
-  /** Minimum iterations before killing a bad basin run. */
-  private static final int BAD_BASIN_MIN_ITERS = 300;
-
   /** Number of refinement passes with decreasing sigma. */
   private static final int REFINEMENT_PASSES = 3;
 
   /** Sigma reduction factor per refinement pass. */
   private static final double[] REFINEMENT_SIGMA_FACTORS = {0.1, 0.03, 0.01};
+
+  private final CMAESRunExecutor executor;
 
   public CMAESTrajectoryOptimizer(TrajectoryProblem problem, int maxEvaluations) {
     this(problem, maxEvaluations, 4, 1e-6, 1e-4, 1e-6);
@@ -69,9 +54,7 @@ public class CMAESTrajectoryOptimizer implements TrajectoryOptimizer {
     this.problem = problem;
     this.maxEvaluations = maxEvaluations;
     this.numExplorationRuns = numExplorationRuns;
-    this.stopFitness = stopFitness;
-    this.relativeTolerance = relativeTolerance;
-    this.absoluteTolerance = absoluteTolerance;
+    this.executor = new CMAESRunExecutor(problem, stopFitness, absoluteTolerance, relativeTolerance);
   }
 
   @Override
@@ -109,7 +92,7 @@ public class CMAESTrajectoryOptimizer implements TrajectoryOptimizer {
         populationSize = basePopSize;
       } else if (globalBestCost < GOOD_BASIN_THRESHOLD) {
         // Good basin found — explore locally around it
-        startPoint = perturbLocal(globalBestVars, sigma, 0.3, run);
+        startPoint = perturbLocal(globalBestVars, sigma, 0.3);
         runSigma = scaleSigma(sigma, 0.5);
         populationSize = basePopSize;
       } else {
@@ -127,20 +110,20 @@ public class CMAESTrajectoryOptimizer implements TrajectoryOptimizer {
       int runBudget = FastMath.min(evalsPerExploration, remainingEvals);
 
       try {
-        RunResult result = runSingleCMAES(startPoint, runSigma, populationSize, runBudget, true);
-        totalEvaluations += result.evaluations;
-        remainingEvals -= result.evaluations;
+        CMAESRunExecutor.RunResult result = executor.execute(startPoint, runSigma, populationSize, runBudget, true);
+        totalEvaluations += result.evaluations();
+        remainingEvals -= result.evaluations();
 
         logger.info(
             "Exploration {}/{}: cost={}, evals={}",
             run + 1,
             numExplorationRuns,
-            result.bestCost,
-            result.evaluations);
+            result.bestCost(),
+            result.evaluations());
 
-        if (result.bestCost < globalBestCost) {
-          globalBestCost = result.bestCost;
-          globalBestVars = result.bestVariables.clone();
+        if (result.bestCost() < globalBestCost) {
+          globalBestCost = result.bestCost();
+          globalBestVars = result.bestVariables().clone();
         }
 
         if (globalBestCost <= problem.getAcceptableCost()) {
@@ -181,22 +164,22 @@ public class CMAESTrajectoryOptimizer implements TrajectoryOptimizer {
         int budget = FastMath.min(refinePassBudget, remainingEvals);
 
         try {
-          RunResult result =
-              runSingleCMAES(globalBestVars.clone(), refineSigma, basePopSize, budget, false);
-          totalEvaluations += result.evaluations;
-          remainingEvals -= result.evaluations;
+          CMAESRunExecutor.RunResult result =
+              executor.execute(globalBestVars.clone(), refineSigma, basePopSize, budget, false);
+          totalEvaluations += result.evaluations();
+          remainingEvals -= result.evaluations();
 
           logger.info(
               "Refinement {}/{} (sigma×{}): cost={}, evals={}",
               pass + 1,
               REFINEMENT_PASSES,
               sigmaFactor,
-              result.bestCost,
-              result.evaluations);
+              result.bestCost(),
+              result.evaluations());
 
-          if (result.bestCost < globalBestCost) {
-            globalBestCost = result.bestCost;
-            globalBestVars = result.bestVariables.clone();
+          if (result.bestCost() < globalBestCost) {
+            globalBestCost = result.bestCost();
+            globalBestVars = result.bestVariables().clone();
           }
         } catch (Exception e) {
           logger.warn("Refinement pass {} failed: {}", pass + 1, e.getMessage());
@@ -221,7 +204,7 @@ public class CMAESTrajectoryOptimizer implements TrajectoryOptimizer {
   // ══════════════════════════════════════════════════════════════════════
 
   /** Small perturbation around a known good point. */
-  private double[] perturbLocal(double[] base, double[] sigma, double scale, int seed) {
+  private double[] perturbLocal(double[] base, double[] sigma, double scale) {
     double[] result = new double[base.length];
     for (int i = 0; i < base.length; i++) {
       result[i] = base[i] + rng.nextGaussian() * sigma[i] * scale;
@@ -256,100 +239,5 @@ public class CMAESTrajectoryOptimizer implements TrajectoryOptimizer {
     for (int i = 0; i < point.length; i++) {
       point[i] = FastMath.max(lower[i], FastMath.min(upper[i], point[i]));
     }
-  }
-
-  // ══════════════════════════════════════════════════════════════════════
-  // Core CMA-ES execution
-  // ══════════════════════════════════════════════════════════════════════
-
-  private record RunResult(double[] bestVariables, double bestCost, int evaluations) {}
-
-  /**
-   * @param earlyKill if true, the convergence checker will kill runs stuck in bad basins
-   */
-  private RunResult runSingleCMAES(
-      double[] startPoint, double[] sigma, int populationSize, int maxEvals, boolean earlyKill) {
-
-    double[] runBestVars = startPoint.clone();
-    double[] runBestCostHolder = {Double.MAX_VALUE};
-
-    MultivariateFunction objectiveFunction =
-        candidate -> {
-          try {
-            SpacecraftState state = problem.propagate(candidate);
-            double cost = problem.computeCost(state);
-            if (Double.isNaN(cost) || Double.isInfinite(cost)) {
-              return EXCEPTION_PENALTY_COST;
-            }
-            synchronized (runBestCostHolder) {
-              if (cost < runBestCostHolder[0]) {
-                runBestCostHolder[0] = cost;
-                System.arraycopy(candidate, 0, runBestVars, 0, candidate.length);
-              }
-            }
-            return cost;
-          } catch (Exception e) {
-            synchronized (runBestCostHolder) {
-              if (EXCEPTION_PENALTY_COST < runBestCostHolder[0]) {
-                runBestCostHolder[0] = EXCEPTION_PENALTY_COST;
-                System.arraycopy(candidate, 0, runBestVars, 0, candidate.length);
-              }
-            }
-            return EXCEPTION_PENALTY_COST;
-          }
-        };
-
-    CMAESOptimizer optimizer = buildOptimizer(maxEvals, earlyKill);
-
-    try {
-      optimizer.optimize(
-          new MaxEval(maxEvals),
-          new ObjectiveFunction(objectiveFunction),
-          GoalType.MINIMIZE,
-          new InitialGuess(startPoint),
-          new CMAESOptimizer.Sigma(sigma),
-          new CMAESOptimizer.PopulationSize(populationSize),
-          new SimpleBounds(problem.getLowerBounds(), problem.getUpperBounds()));
-    } catch (org.hipparchus.exception.MathRuntimeException e) {
-      // Budget exhausted (TooManyEvaluationsException) or numerical issue — use best found so far
-      logger.debug("CMA-ES run ended early: {}", e.getMessage());
-    }
-
-    return new RunResult(runBestVars, runBestCostHolder[0], optimizer.getEvaluations());
-  }
-
-  private @NonNull CMAESOptimizer buildOptimizer(int maxEvals, boolean earlyKill) {
-    ConvergenceChecker<PointValuePair> checker =
-        new ConvergenceChecker<>() {
-          private int iterationCount = 0;
-
-          @Override
-          public boolean converged(int iteration, PointValuePair previous, PointValuePair current) {
-            iterationCount++;
-
-            // Never converge too early
-            if (iterationCount < 100) return false;
-
-            // ── Early kill: abort runs stuck in bad basins ──
-            if (earlyKill
-                && iterationCount > BAD_BASIN_MIN_ITERS
-                && current.getValue() > BAD_BASIN_KILL_THRESHOLD) {
-              return true;
-            }
-
-            // Don't converge prematurely if we haven't reached acceptable cost yet
-            if (current.getValue() > problem.getAcceptableCost() && iterationCount < 500) {
-              return false;
-            }
-
-            // Standard convergence: cost stopped improving
-            double diff = FastMath.abs(previous.getValue() - current.getValue());
-            return diff <= absoluteTolerance
-                || diff <= relativeTolerance * FastMath.abs(current.getValue());
-          }
-        };
-
-    return new CMAESOptimizer(
-        maxEvals, stopFitness, true, 0, 0, new MersenneTwister(), false, checker);
   }
 }
